+++ hikyuu_cpp/hikyuu/trade_manage/OrderBrokerBase.cpp
@@ -26,7 +26,7 @@ OrderBrokerBase::OrderBrokerBase(const string& name) : m_name(name) {}
 OrderBrokerBase::~OrderBrokerBase() {}
 
 Datetime OrderBrokerBase::buy(Datetime datetime, const string& market, const string& code,
-                              price_t price, int num) {
+                              price_t price, double num) {
     Datetime tradetime;
     try {
         tradetime = _buy(datetime, market, code, price, num);
@@ -38,7 +38,7 @@ Datetime OrderBrokerBase::buy(Datetime datetime, const string& market, const str
 }
 
 Datetime OrderBrokerBase::sell(Datetime datetime, const string& market, const string& code,
-                               price_t price, int num) {
+                               price_t price, double num) {
     Datetime tradetime;
     try {
         tradetime = _sell(datetime, market, code, price, num);
