+++ hikyuu_cpp/hikyuu/trade_manage/OrderBrokerBase.h
@@ -52,7 +52,7 @@ public:
      * @return 操作执行的时刻。实盘时，应返回委托单时间或服务器交易时间。
      */
     Datetime buy(Datetime datetime, const string& market, const string& code, price_t price,
-                 int num);
+                 double num);
 
     /**
      * 执行卖出操作
@@ -64,7 +64,7 @@ public:
      * @return 操作执行的时刻。实盘时，应返回委托单时间或服务器交易时间。
      */
     Datetime sell(Datetime datetime, const string& market, const string& code, price_t price,
-                  int num);
+                  double num);
 
     /**
      * 执行实际买入操作
@@ -76,7 +76,7 @@ public:
      * @return 操作执行的时刻。实盘时，应返回委托单时间或服务器交易时间。
      */
     virtual Datetime _buy(Datetime datetime, const string& market, const string& code,
-                          price_t price, int num) = 0;
+                          price_t price, double num) = 0;
 
     /**
      * 执行实际卖出操作
@@ -88,7 +88,7 @@ public:
      * @return 操作执行的时刻。实盘时，应返回委托单时间或服务器交易时间。
      */
     virtual Datetime _sell(Datetime datetime, const string& market, const string& code,
-                           price_t price, int num) = 0;
+                           price_t price, double num) = 0;
 
 protected:
     string m_name;
