+++ hikyuu_cpp/hikyuu/trade_manage/PositionRecord.h
@@ -26,7 +26,7 @@ class HKU_API PositionRecord {
 public:
     PositionRecord();
     PositionRecord(const Stock& stock, const Datetime& takeDatetime, const Datetime& cleanDatetime,
-                   size_t number, price_t stoploss, price_t goalPrice, size_t totalNumber,
+                   double number, price_t stoploss, price_t goalPrice, double totalNumber,
                    price_t buyMoney, price_t totalCost, price_t totalRisk, price_t sellMoney);
 
     /** 仅用于python的__str__ */
@@ -35,10 +35,10 @@ public:
     Stock stock;             ///< 交易对象
     Datetime takeDatetime;   ///< 初次建仓日期
     Datetime cleanDatetime;  ///< 平仓日期，当前持仓记录中为Null<Datetime>()
-    size_t number;           ///< 当前持仓数量
+    double number;           ///< 当前持仓数量
     price_t stoploss;        ///< 当前止损价
     price_t goalPrice;       ///< 当前的目标价格
-    size_t totalNumber;      ///< 累计持仓数量
+    double totalNumber;      ///< 累计持仓数量
     price_t buyMoney;        ///< 累计买入资金
     price_t totalCost;       ///< 累计交易总成本
     price_t totalRisk;  ///< 累计交易风险 = 各次 （买入价格-止损)*买入数量, 不包含交易成本
