+++ hikyuu_cpp/hikyuu/trade_manage/imp/FixedA2015TradeCost.h
@@ -29,7 +29,7 @@ public:
      * @return CostRecord 交易成本记录
      */
     virtual CostRecord getBuyCost(const Datetime& datetime, const Stock& stock, price_t price,
-                                  size_t num) const override;
+                                  double num) const override;
 
     /**
      * 计算卖出成本
@@ -40,7 +40,7 @@ public:
      * @return CostRecord 交易成本记录
      */
     virtual CostRecord getSellCost(const Datetime& datetime, const Stock& stock, price_t price,
-                                   size_t num) const override;
+                                   double num) const override;
 
     /** 子类私有变量克隆接口 */
     virtual TradeCostPtr _clone() override;
