+++ hikyuu_cpp/hikyuu/trade_manage/imp/ZeroTradeCost.cpp
@@ -14,12 +14,12 @@ ZeroTradeCost::ZeroTradeCost() : TradeCostBase("TC_Zero") {}
 ZeroTradeCost::~ZeroTradeCost() {}
 
 CostRecord ZeroTradeCost ::getBuyCost(const Datetime& datetime, const Stock& stock, price_t price,
-                                      size_t num) const {
+                                      double num) const {
     return CostRecord();
 }
 
 CostRecord ZeroTradeCost ::getSellCost(const Datetime& datetime, const Stock& stock, price_t price,
-                                       size_t num) const {
+                                       double num) const {
     return CostRecord();
 }
